Change how "minify" is used throughout production build
It used to be a trigger to enable production build features such as `HashedModuleIdsPlugin` and `mode: "production"`. However, there are cases you need to enable those features without minifying the asset. For this reason, now `minify` is used only for enabling minification and `context.stageIs("production")` is used whenever we need to control optimization features.